pyproject.toml
[project]
name = "ai-memory"
version = "0.1.0"
description = ""
authors = [
    {name = "echoedinvoker",email = "echoedinvoker@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "langchain-anthropic (>=0.3.9,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
